gateway/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Redis
redis>=5.0.0

# AWS/MinIO client
aiobotocore>=2.9.0

# Docker
aiodocker>=0.21.0

# JWT Authentication
PyJWT>=2.8.0

# Prometheus metrics
prometheus-client>=0.19.0

# Additional utilities
python-multipart>=0.0.6
httpx>=0.26.0
